Use double instead of float (cherry picked from commit c90c5d503543a960a05b43752a5dff9ccf4bcd30)

+++ writerperfect/source/filter/DocumentCollector.hxx
@@ -204,7 +204,7 @@ private:
 
     // section styles
     std::vector<SectionStyle *> mSectionStyles;
-    float mfSectionSpaceAfter;
+    double mfSectionSpaceAfter;
 
     // table styles
     std::vector<TableStyle *> mTableStyles;
